The AuthenticationController is now refactored to AuthController and changed it's content to include PostMapping for /register.
The CustomUserDetailsService is now refactored to authService to more accurately describe its usages and functions and now includes a registerUser-method to properly test the function of registering a user.

The above mentioned register-function AND the previously existing login now uses DTO's to properly converse with JSON to the frontend.

UserController now dont have much content. It's usage has been moved to AuthController to divide the UserController to handle its own UserDetails.

The references to these in SecurityConfig are also refactored.